Make non-repudiation integration exclusive to enterprise edition
Closes #9182

changelog_begin
changelog_end
