+++ daml-script/test/BUILD.bazel
@@ -182,7 +182,8 @@ da_scala_test_suite(
         "//language-support/scala/bindings-akka",
         "//ledger-api/rs-grpc-bridge",
         "//ledger-api/testing-utils",
-        "//ledger-service/http-json",
+        "//ledger-service/http-json:http-json-ce",
+        "//ledger-service/http-json-cli:ce",
         "//ledger-service/jwt",
         "//ledger-service/lf-value-json",
         "//ledger/caching",
