+++ ledger-service/http-json-cli/BUILD.bazel
@@ -1,0 +1,47 @@
+# Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load(
+    "//bazel_tools:scala.bzl",
+    "da_scala_library",
+)
+
+da_scala_library(
+    name = "base",
+    srcs = glob(["src/main/scala/**/*.scala"]),
+    scala_deps = [
+        "@maven//:com_github_scopt_scopt",
+        "@maven//:org_scalaz_scalaz_core",
+        "@maven//:com_typesafe_akka_akka_stream",
+        "@maven//:com_typesafe_scala_logging_scala_logging",
+    ],
+    visibility = ["//ledger-service:__subpackages__"],
+    deps = [
+        "//ledger-service/cli-opts",
+        "//ledger-service/utils",
+        "//ledger/ledger-api-common",
+        "@maven//:io_netty_netty_handler",
+        "@maven//:org_slf4j_slf4j_api",
+    ],
+)
+
+[
+    da_scala_library(
+        name = edition,
+        srcs = glob(["{}/src/main/scala/**/*.scala".format(edition)]),
+        scala_deps = [
+            "@maven//:com_github_scopt_scopt",
+            "@maven//:com_typesafe_scala_logging_scala_logging",
+        ],
+        visibility = ["//visibility:public"],
+        exports = [":base"],
+        deps = [
+            ":base",
+            "//ledger/ledger-api-common",
+        ],
+    )
+    for edition in [
+        "ce",
+        "ee",
+    ]
+]
