+++ ledger-service/http-json-cli/ee/src/main/scala/com/daml/http/Cli.scala
@@ -1,0 +1,9 @@
+// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.http
+
+object Cli extends CliBase {
+  override protected def configParser(getEnvVar: String => Option[String]): OptionParser =
+    new OptionParser(getEnvVar) with NonRepudiationOptions
+}
