+++ ledger-service/http-json-oracle/BUILD.bazel
@@ -70,9 +70,9 @@ da_scala_test(
         "//daml-lf/transaction",
         "//language-support/scala/bindings-akka",
         "//ledger-api/rs-grpc-bridge",
-        "//ledger-service/http-json",
-        "//ledger-service/http-json:integration-tests-lib",
-        "//ledger-service/http-json-testing",
+        "//ledger-service/http-json:http-json-ee",
+        "//ledger-service/http-json:integration-tests-lib-ee",
+        "//ledger-service/http-json-testing:ee",
         "//ledger-service/jwt",
         "//ledger-service/utils",
         "//ledger/ledger-api-common",
