+++ ledger-service/http-json-testing/BUILD.bazel
@@ -14,50 +14,57 @@ hj_scalacopts = lf_scalacopts + [
     "-P:wartremover:traverser:org.wartremover.warts.NonUnitStatements",
 ]
 
-da_scala_library(
-    name = "http-json-testing",
-    srcs = glob(["src/main/scala/**/*.scala"]),
-    plugins = [
-        "@maven//:org_typelevel_kind_projector_{}".format(scala_version_suffix),
-        silencer_plugin,
-    ],
-    scala_deps = [
-        "@maven//:com_typesafe_akka_akka_actor",
-        "@maven//:com_typesafe_akka_akka_http_core",
-        "@maven//:com_typesafe_akka_akka_stream",
-        "@maven//:io_spray_spray_json",
-        "@maven//:org_scalacheck_scalacheck",
-        "@maven//:org_scalactic_scalactic",
-        "@maven//:org_scalatest_scalatest",
-        "@maven//:org_tpolecat_doobie_core",
-        "@maven//:org_tpolecat_doobie_free",
-        "@maven//:org_typelevel_cats_core",
-        "@maven//:org_typelevel_cats_effect",
-        "@maven//:org_typelevel_cats_free",
-    ],
-    scalacopts = hj_scalacopts + [
-        "-P:silencer:lineContentFilters=import scala.collection.compat",
-    ],
-    tags = ["maven_coordinates=com.daml:http-json-testing:__VERSION__"],
-    visibility = ["//visibility:public"],
-    runtime_deps = [
-        "@maven//:ch_qos_logback_logback_classic",
-    ],
-    deps = [
-        "//bazel_tools/runfiles:scala_runfiles",
-        "//language-support/scala/bindings-akka",
-        "//ledger-api/rs-grpc-bridge",
-        "//ledger-service/http-json",
-        "//ledger-service/jwt",
-        "//ledger-service/utils",
-        "//ledger/caching",
-        "//ledger/ledger-api-auth",
-        "//ledger/ledger-api-common",
-        "//ledger/participant-integration-api",
-        "//ledger/participant-state",
-        "//ledger/sandbox-classic",
-        "//ledger/sandbox-common",
-        "//libs-scala/auth-utils",
-        "//libs-scala/ports",
-    ],
-)
+[
+    da_scala_library(
+        name = "{}".format(edition),
+        srcs = glob(["src/main/scala/**/*.scala"]),
+        plugins = [
+            "@maven//:org_typelevel_kind_projector_{}".format(scala_version_suffix),
+            silencer_plugin,
+        ],
+        scala_deps = [
+            "@maven//:com_typesafe_akka_akka_actor",
+            "@maven//:com_typesafe_akka_akka_http_core",
+            "@maven//:com_typesafe_akka_akka_stream",
+            "@maven//:io_spray_spray_json",
+            "@maven//:org_scalacheck_scalacheck",
+            "@maven//:org_scalactic_scalactic",
+            "@maven//:org_scalatest_scalatest",
+            "@maven//:org_tpolecat_doobie_core",
+            "@maven//:org_tpolecat_doobie_free",
+            "@maven//:org_typelevel_cats_core",
+            "@maven//:org_typelevel_cats_effect",
+            "@maven//:org_typelevel_cats_free",
+        ],
+        scalacopts = hj_scalacopts + [
+            "-P:silencer:lineContentFilters=import scala.collection.compat",
+        ],
+        tags = ["maven_coordinates=com.daml:http-json-testing:__VERSION__"],
+        visibility = ["//visibility:public"],
+        runtime_deps = [
+            "@maven//:ch_qos_logback_logback_classic",
+        ],
+        deps = [
+            "//bazel_tools/runfiles:scala_runfiles",
+            "//language-support/scala/bindings-akka",
+            "//ledger-api/rs-grpc-bridge",
+            "//ledger-service/http-json:http-json-{}".format(edition),
+            "//ledger-service/http-json-cli:{}".format(edition),
+            "//ledger-service/jwt",
+            "//ledger-service/utils",
+            "//ledger/caching",
+            "//ledger/ledger-api-auth",
+            "//ledger/ledger-api-common",
+            "//ledger/participant-integration-api",
+            "//ledger/participant-state",
+            "//ledger/sandbox-classic",
+            "//ledger/sandbox-common",
+            "//libs-scala/auth-utils",
+            "//libs-scala/ports",
+        ],
+    )
+    for edition in [
+        "ce",
+        "ee",
+    ]
+]
