+++ ledger-service/http-json/src/failure/scala/http/HttpTestFixture.scala
@@ -6,7 +6,7 @@ package com.daml.http
 import akka.http.scaladsl.model.Uri
 import com.daml.bazeltools.BazelRunfiles
 import com.daml.http.json.{DomainJsonDecoder, DomainJsonEncoder}
-import com.daml.ledger.client.LedgerClient
+import com.daml.ledger.client.{LedgerClient => DamlLedgerClient}
 import com.daml.ports.LockedFreePort
 import com.daml.testing.postgresql.PostgresAroundAll
 import java.net.InetAddress
@@ -47,13 +47,14 @@ trait HttpFailureTestFixture extends ToxicSandboxFixture with PostgresAroundAll 
       BazelRunfiles.requiredResource("ledger-service/http-json/Account.dar"),
     )
 
-  protected def allocateParty(client: LedgerClient, displayName: String): Future[domain.Party] =
+  protected def allocateParty(client: DamlLedgerClient, displayName: String): Future[domain.Party] =
     client.partyManagementClient
       .allocateParty(None, Some(displayName), None)
       .map(p => domain.Party(p.party))
 
-  protected def withHttpService[A]
-      : ((Uri, DomainJsonEncoder, DomainJsonDecoder, LedgerClient) => Future[A]) => Future[A] = {
+  protected def withHttpService[A]: (
+      (Uri, DomainJsonEncoder, DomainJsonDecoder, DamlLedgerClient) => Future[A]
+  ) => Future[A] = {
     println(proxy.getUpstream())
     HttpServiceTestFixture.withHttpService(
       this.getClass.getSimpleName,
